test-requirements.txt
nose>=1.2.1,<2.0
spur.local>=0.3.7,<0.4
tempman>=0.1.2,<0.2
